# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Spottruck API
  description: |
    Plataforma SaaS de subastas inversas para transporte de carga en Argentina.

    ## Autenticación
    Todas las rutas protegidas requieren `Authorization: Bearer <accessToken>`.

    ## Paginación
    Los endpoints de lista usan paginación con `page` y `limit`:
    - `page`: número de página (default 1)
    - `limit`: items por página (default 20, max 100)

    ## Códigos de error
    | Código | Descripción |
    |--------|-------------|
    | 400 | Bad Request — datos inválidos |
    | 401 | Unauthorized — token inválido o expirado |
    | 403 | Forbidden — sin permisos para este recurso |
    | 404 | Not Found — recurso no existe |
    | 409 | Conflict — estado incompatible (ej: auction ya cerrada) |
    | 500 | Internal Server Error |
  version: 1.0.0
  contact:
    name: Spottruck Dev Team
    email: dev@spottruck.com
servers:
  - url: http://localhost:4000
    description: Development
  - url: https://api.spottruck.com
    description: Production
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar nuevo usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: empresa@ejemplo.com
              password: MiContraseña123!
              role: COMPANY
              companyName: Acme Logistica SRL
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refrescar tokens de acceso
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Nuevos tokens generados
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Cerrar sesión
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Sesión cerrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
  /auth/me:
    get:
      tags: [Auth]
      summary: Obtener usuario autenticado
      responses:
        '200':
          description: Usuario actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ========================
  # USERS
  # ========================
...
  /auctions:
    get:
      tags: [Auctions]
      summary: Listar auctions
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AuctionStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AuctionType'
        - name: tripId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lista de auctions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Auction'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /auctions/{id}:
    get:
      tags: [Auctions]
      summary: Obtener auction con bids
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Auction con bids
          content:
            application/json:
              schema:
                type: object
                properties:
                  auction:
                    $ref: '#/components/schemas/Auction'
                  bids:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bid'
        '404':
          $ref: '#/components/responses/NotFound'
  /auctions/{id}/bid:
    post:
      tags: [Auctions]
      summary: Realizar una oferta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  description: Monto de la oferta (debe ser menor al precio actual — subasta inversa)
                  example: 35000
      responses:
        '201':
          description: Oferta realizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    $ref: '#/components/schemas/Bid'
                  auction:
                    $ref: '#/components/schemas/Auction'
        '400':
          description: Oferta inválida (monto mayor/igual al actual, auction cerrada)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Solo drivers pueden pujar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auctions/{id}/bids:
    get:
      tags: [Auctions]
      summary: Obtener historial de ofertas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de bids
          content:
            application/json:
              schema:
                type: object
                properties:
                  bids:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bid'

  # ========================
  # PAYMENTS
  # ========================
  /payments/{tripId}:
    get:
      tags: [Payments]
      summary: Obtener pago asociado a un viaje
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/{tripId}/release:
    post:
      tags: [Payments]
      summary: Liberar pago (empresa confirma entrega)
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pago liberado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Pago no puede ser liberado (estado inválido)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de acceso (15 min de TTL)
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Token inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sin permisos para este recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    UserRole:
      type: string
      enum: [COMPANY, DRIVER, ADMIN]
...
    AuctionType:
      type: string
      enum: [OPEN, DUTCH, SEALED]
    AuctionStatus:
      type: string
      enum: [PENDING, OPEN, CLOSED, SETTLED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [PENDING, HELD, RELEASED, REFUNDED, DISPUTED]

    # ---- Shared ----
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        code:
          type: string
          example: VALIDATION_ERROR
        details:
          type: object
          nullable: true
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 142
        totalPages:
          type: integer
          example: 8

    # ---- Auth ----
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: empresa@ejemplo.com
        password:
          type: string
          minLength: 8
          example: MiContraseña123!
        role:
          $ref: '#/components/schemas/UserRole'
        companyName:
          type: string
          description: Requerido si role=COMPANY
          example: Acme Logística SRL
        driverLicense:
          type: string
          description: Requerido si role=DRIVER
          example: D1234567
        vehiclePlate:
          type: string
          description: Opcional para DRIVER
          example: AB123CD
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT de acceso (válido 15 min)
        refreshToken:
          type: string
          description: JWT de refresh (válido 7 días)
        user:
          $ref: '#/components/schemas/User'

    # ---- User ----
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        companyName:
          type: string
          nullable: true
        driverLicense:
          type: string
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        ratingAvg:
          type: number
          nullable: true
          example: 4.7
        tripsCompleted:
          type: integer
          default: 0
        trustScore:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Auction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/AuctionType'
        status:
          $ref: '#/components/schemas/AuctionStatus'
        currentPrice:
          type: number
          description: Precio actual (en ARS)
        reservePrice:
          type: number
          nullable: true
          description: Precio mínimo de reserva (opcional)
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        extensionCount:
          type: integer
          default: 0
          description: Cantidad de extensiones por anti-sniping
        winnerId:
          type: string
          format: uuid
          nullable: true
        bidsCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
    Bid:
      type: object
      properties:
        id:
          type: string
          format: uuid
        auctionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: number
          description: Monto ofertado (menor = mejor en subasta inversa)
        createdAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
            companyName:
              type: string
              nullable: true
            ratingAvg:
              type: number
              nullable: true

    # ---- Payment ----
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        driverId:
          type: string
          format: uuid
        amount:
          type: number
        platformFee:
          type: number
          description: Fee de plataforma (8% del amount)
        netAmount:
          type: number
          description: Monto neto para el conductor (amount - platformFee)
        status:
          $ref: '#/components/schemas/PaymentStatus'
        holdExpiresAt:
          type: string
          format: date-time
        releasedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ---- Rating ----
...